Rename constants.dart as pkg.dart
