+++ lib/widgets/checkbox_group.dart
@@ -29,7 +29,7 @@
 /// Authors: Tony Chen
 
 import 'package:flutter/material.dart';
-import 'package:markdown_widgets/constants/constants.dart'
+import 'package:markdown_widgets/constants/pkg.dart'
     show contentWidthFactor;
 
 class CheckboxGroup extends StatefulWidget {
